requirements-gpu.txt
tensorflow-gpu==1.7.0
scipy==1.0.0
scikit_image==0.13.1
opencv_python==3.4.1.15
numpy==1.14.0
tqdm==4.23.4
skimage==0.0
